+++ app/Controller/AdminController/ColorController.php
@@ -1,7 +1,10 @@
 <?php
+
 namespace App\Controller\AdminController;
+
 use App\Models\Color;
 use App\Models\ProductVariant;
+
 class ColorController
 {
     public function index()
@@ -42,18 +45,24 @@ class ColorController
         // Validate data
         if (empty($data['name'])) {
             $_SESSION['error'] = 'Tên màu không được để trống!';
-            return redirect('/admin/colors/edit/'.$id);
+            return redirect('/admin/colors/edit/' . $id);
         } else {
             // Update color
             Color::update($data, $id);
             $_SESSION['message'] = 'Cập nhật màu sắc thành công!';
-            return redirect('/admin/colors/edit/'.$id);
+            return redirect('/admin/colors/edit/' . $id);
         }
     }
 
     public function destroy($id)
     {
-        ProductVariant::deleteByColorId($id);
+        $variants = ProductVariant::select(['product_variant.*'])
+            ->where('id_color', '=', $id)
+            ->get();
+        if (count($variants) > 0) {
+            $_SESSION['error'] = 'Không thể xóa màu sắc này vì nó đang được sử dụng trong các sản phẩm!';
+            return redirect('admin/colors');
+        }
         Color::delete($id);
         $_SESSION['confim'] = 'Xóa màu sắc thành công!';
         return redirect('/admin/colors');
